separates indexer to own provider
